Updated CSS for tutorial widgets
