--- scripts/tutorial-cards.css
@@ -1,120 +1,0 @@
-.tutorial-card {
-    background-color: var(--background-color);
-    border: 1px solid var(--spectrum-gray-200);
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-    border-radius: 5px;
-    word-break: break-word;
-  }
-  .tutorial-card:hover {
-    box-shadow: var(--box-shadow-2);
-  }
-  .tutorial-card * {
-    box-sizing: border-box;
-  }
-  .tutorial-card > * {
-    flex-grow: 1;
-    margin-bottom: 0;
-  }
-  .tutorial-card .tutorial-card-banner {
-    position: absolute;
-  top: 32px;
-  left: 0;
-  background-color: var(--non-spectrum-light-coral);
-  color: var(--background-color);
-  padding: 8px 16px;
-  border-radius: 0 4px 4px 0;
-  text-transform: uppercase;
-  font-size: var(--spectrum-font-size-75);
-  line-height: 17px;
-  font-weight: var(--font-weight-700);
-  letter-spacing: 0.72px;
-  margin: 0;
-  }
-  .tutorial-card .tutorial-card-tag-text {
-    display: flex;
-  margin: 0 0 8px;
-  }
-  .tutorial-card .tutorial-card-title-text {
-    text-transform: uppercase;
-  margin: 0;
-  font-size: var(--spectrum-font-size-100);
-  font-weight: var(--font-weight-400);
-  line-height: 21px;
-  color: var(--spectrum-gray-700);
-  }
-  .tutorial-card .tutorial-card-event-info {
-    display: flex;
-  gap: 4px;
-  margin: 8px 0;
-  line-height: 0;
-  }
-  .tutorial-card .tutorial-card-event-time {
-    display: flex;
-    flex-direction: column;
-    gap: 2px;
-  }
-  .tutorial-card .tutorial-card-description-text,
-  .tutorial-card .tutorial-card-description-text * {
-    margin: 0;
-  font-size: var(--spectrum-font-size-100);
-  line-height: 21px;
-  color: var(--spectrum-gray-700);
-  }
-  .tutorial-card .tutorial-card-figure {
-    max-height: 200px;
-  min-height: 72px;
-  overflow: hidden;
-  position: relative;
-  }
-  .tutorial-card .tutorial-card-content {
-    padding: 16px 16px 12px 15px;
-  min-height: 190px;
-  }
-  .tutorial-card .tutorial-card-meta-info {
-    display: flex;
-  gap: 12px;
-  margin: 8px 0;
-  }
-  .tutorial-card .tutorial-card-author-info {
-    display: flex;
-  margin-top: 16px;
-  flex-wrap: wrap;
-  gap: 10px 0;
-  }
-  .tutorial-card .tutorial-card-footer {
-    padding: 0 16px 22px 15px;
-    display: flex;
-    align-items: flex-end;
-    justify-content: space-between;
-    width: 100%;
-  }
-  .tutorial-card .tutorial-card-footer .tutorial-card-options {
-    position: relative;
-    display: flex;
-    top: 8px;
-  }
-  .tutorial-card .tutorial-card-footer .tutorial-card-cta-element {
-    background-color: var(--background-color);
-    color: var(--non-spectrum-dark-navy-blue);
-    margin: 0;
-    display: flex;
-    align-items: center;
-    gap: 4px;
-    top: 4px;
-    position: relative;
-    font-weight: var(--font-weight-600);
-    font-size: var(--spectrum-font-size-100);
-    line-height: 17px;
-  }
-  .tutorial-card .tutorial-card-meta-anchor {
-    margin: 0;
-    display: flex;
-    flex-direction: row-reverse;
-    align-items: center;
-    gap: 4px;
-    color: var(--non-spectrum-grey-updated);
-    text-decoration: none;
-    font-size: var(--spectrum-font-size-75);
-  }
