+++ scripts/tutorial-widgets/tutorial-widgets.css
@@ -1,0 +1,68 @@
+.tutorial-widgets {
+  display: flex;
+  flex-wrap: nowrap;
+  gap: 16px;
+  overflow-x: auto;
+  overflow-y: hidden;
+  padding-bottom: 2px;
+  min-height: 400px;
+  align-items: stretch;
+}
+
+.tutorial-widgets > div {
+  flex: 0 0 264px;
+}
+
+.tutorial-widgets a {
+  font-size: 18px;
+  line-height: 17px;
+  margin: 2px 0 0 0;
+}
+
+.tutorial-widgets > div:first-child .bookmark .exl-tooltip-label,
+.tutorial-widgets > div:first-child .copy-link .exl-tooltip-label {
+left: 100%;
+transform: translateX(-20%);
+}
+
+.tutorial-widgets > div:first-child .bookmark .exl-tooltip-label::after,
+.tutorial-widgets > div:first-child .copy-link .exl-tooltip-label::after {
+left: 10%;
+}
+
+.tutorial-widgets > div:first-child .bookmark + .copy-link .exl-tooltip-label {
+left: 50%;
+transform: translateX(-50%);
+}
+
+.tutorial-widgets > div:first-child .bookmark + .copy-link .exl-tooltip-label::after {
+left: 50%;
+}
+
+@media only screen and (min-width: 600px) {
+  .tutorial-widgets {
+      gap: 25px;
+      overflow-x: auto;
+      overflow-y: hidden;
+  }
+  .tutorial-widgets > div {
+      flex: 0 0 286px;
+  } 
+}
+
+@media only screen and (min-width: 900px) {
+  .tutorial-widgets {
+      gap: 32px;
+  }
+  .tutorial-widgets > div {
+      flex: 0 0 256px;
+  }
+}
+
+@media only screen and (min-width: 1200px) {
+  .tutorial-widgets {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+  }
+}
